requirements.txt
torch
torchvision
transformers
pandas
numpy
Pillow
requests
scikit-learn
tqdm
matplotlib
seaborn
datasets